fixed where we create tile and layer GPUBuffer data; added greyscale support for WebGPU

+++ s2/workers/process/line.ts
@@ -202,7 +202,7 @@ export default class LineWorker extends VectorWorker implements LineWorkerSpec {
     }
 
     // Upon building the batches, convert to buffers and ship.
-    const vertexBuffer = new Int16Array(vertices).buffer
+    const vertexBuffer = new Float32Array(vertices).buffer
     const lengthSoFarBuffer = new Float32Array(lengthSoFar).buffer
     const featureGuideBuffer = new Float32Array(featureGuide).buffer
     // ship the vector data.
